Failing tests for sparse array builder
Also renamed Enumerable namespace as it clashes wih the Linq type
